# packages/contracts/openapi.yaml
openapi: 3.0.1
info:
  title: CulturaZ API
  description: API REST do marketplace CulturaZ.
  contact:
    name: Ângelo Oliveira
    email: dev.angelooliveira@gmail.com
  license:
    name: MIT
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/auth/me:
    get:
      tags:
      - auth-controller
      operationId: me_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    UserResponse:
      required:
      - createdAt
      - email
      - id
      - name
      - roles
      - status
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - PENDING_VERIFICATION
          - DELETED
        roles:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - CUSTOMER
            - SELLER
            - ADMIN
            - SUPPORT
        createdAt:
          type: string
          format: date-time
...
    RegisterRequest:
      required:
      - email
      - name
      - password
      type: object
      properties:
        name:
          maxLength: 120
          minLength: 2
          type: string
        email:
          maxLength: 160
          minLength: 0
          type: string
        password:
          maxLength: 100
          minLength: 8
          pattern: "^(?=.*[A-Za-z])(?=.*\\d).+$"
          type: string
        phone:
          maxLength: 20
          minLength: 0
          type: string
    AuthResponse:
      required:
      - accessToken
      - expiresInSeconds
      - refreshToken
      - user
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/UserResponse"
    RefreshTokenRequest:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT